Refactor imports and files structure
