+++ src/App.js
@@ -1,6 +1,6 @@
 import { BrowserRouter, Route, Routes } from 'react-router-dom';
 import { routes } from './routes';
-import Navigation from './components/shared/Navigation/Navigation.jsx';
+import { Navigation } from './components/shared';
 
 function App() {
     return (
@@ -9,6 +9,7 @@ function App() {
                 <Navigation />
                 <Routes>
                     {routes.map((route) => (
+                        // eslint-disable-next-line react/jsx-key
                         <Route {...route} />
                     ))}
                 </Routes>
