+++ src/components/index.js
@@ -1,0 +1,2 @@
+export * from './userProfile/profileInformation/ProfileInformation.jsx';
+export * from './userProfile/profileTabs/ProfileTabs.jsx';
