+++ src/pages/index.js
@@ -1,1 +1,2 @@
 export * from './MainPage/MainPage.jsx';
+export * from './userProfile/UserProfile.jsx';
