+++ src/pages/userProfile/UserProfile.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import ProfileTabs from '../../components/userProfile/profileTabs/ProfileTabs.jsx';
 import ProfileInformation from '../../components/userProfile/profileInformation/ProfileInformation.jsx';
 
-function UserProfile(props) {
+export function UserProfile(props) {
     return (
         <>
             <ProfileInformation />
@@ -10,5 +10,3 @@ function UserProfile(props) {
         </>
     );
 }
-
-export default UserProfile;
